fixed some image resolution
