+++ Data-Generator/generateCardBackground.py
@@ -28,23 +28,34 @@ background_paths = glob.glob(f"{BACKGROUNDS_DIR}/*.jpg") + \
                    glob.glob(f"{BACKGROUNDS_DIR}/*.png") + \
                    glob.glob(f"{BACKGROUNDS_DIR}/*.jpeg")
 
-def load_cards(csv_path):
+def load_cards(csv_paths):
+    """Load cards from one or multiple CSV files"""
     cards = []
-    with open(csv_path, newline='', encoding='utf-8') as csvfile:
-        reader = csv.DictReader(csvfile)
-        for row in reader:
-            img_url = row['Image']
-            img_id = img_url.split('/')[-1].split('.')[0]
-            cards.append({
-                'img_id': img_id,
-                'product_name': row['Product Name'],
-                'image_url': img_url
-            })
+    if isinstance(csv_paths, str):
+        csv_paths = [csv_paths]
+    
+    for csv_path in csv_paths:
+        try:
+            with open(csv_path, newline='', encoding='utf-8') as csvfile:
+                reader = csv.DictReader(csvfile)
+                for row in reader:
+                    img_url = row['Image']
+                    img_id = img_url.split('/')[-1].split('.')[0]
+                    cards.append({
+                        'img_id': img_id,
+                        'product_name': row['Product Name'],
+                        'image_url': img_url
+                    })
+        except FileNotFoundError:
+            print(f"Warning: CSV file not found - {csv_path}")
+            continue
+        except Exception as e:
+            print(f"Error processing {csv_path}: {str(e)}")
+            continue
     return cards
 
 def apply_glare_effect(image):
-    """Add INTENSE artificial glare with dramatic highlights"""
-    # Only apply glare to 80% of cards (more frequent)
+    """Add intense artificial glare with dramatic highlights"""
     if random.random() > 0.8:
         return image
     
@@ -52,13 +63,11 @@ def apply_glare_effect(image):
     overlay = Image.new('RGBA', (width, height), (0, 0, 0, 0))
     draw = ImageDraw.Draw(overlay)
     
-    # DRAMATIC glare parameters
-    intensity = random.uniform(0.5, 1.2)  # Much stronger intensity range
-    cx = random.uniform(-0.1 * width, 1.1 * width)  # More centered
+    intensity = random.uniform(0.5, 1.2)
+    cx = random.uniform(-0.1 * width, 1.1 * width)
     cy = random.uniform(-0.1 * height, 1.1 * height)
-    radius = random.randint(int(min(width, height)*0.3), int(min(width, height)*1.2))  # Larger radius
+    radius = random.randint(int(min(width, height)*0.3), int(min(width, height)*1.2))
     
-    # Create MULTIPLE glare spots (30% chance)
     if random.random() < 0.3:
         for _ in range(random.randint(1, 3)):
             sub_intensity = random.uniform(0.3, 0.8)
@@ -67,84 +76,66 @@ def apply_glare_effect(image):
             sub_radius = random.randint(int(radius*0.3), int(radius*0.7))
             
             for i in range(sub_radius, 0, -1):
-                alpha = int(255 * sub_intensity * (i/sub_radius)**0.3)  # Sharper falloff
+                alpha = int(255 * sub_intensity * (i/sub_radius)**0.3)
                 color = (255, 255, 255, alpha)
-                draw.ellipse(
-                    (sub_cx - i, sub_cy - i, sub_cx + i, sub_cy + i),
-                    fill=color
-                )
+                draw.ellipse((sub_cx - i, sub_cy - i, sub_cx + i, sub_cy + i), fill=color)
     
-    # Main glare spot (always present)
     for i in range(radius, 0, -1):
-        alpha = int(255 * intensity * (i/radius)**0.2)  # Very slow falloff for intense core
+        alpha = int(255 * intensity * (i/radius)**0.2)
         color = random.choice([
-            (255, 255, 255, alpha),  # Bright white
-            (255, 240, 150, alpha),  # Warm golden
-            (180, 220, 255, alpha)   # Cool blue
+            (255, 255, 255, alpha),
+            (255, 240, 150, alpha),
+            (180, 220, 255, alpha)
         ])
-        draw.ellipse(
-            (cx - i, cy - i, cx + i, cy + i),
-            fill=color
-        )
+        draw.ellipse((cx - i, cy - i, cx + i, cy + i), fill=color)
     
-    # Apply blur - less blur for sharper glare
     overlay = overlay.filter(ImageFilter.GaussianBlur(radius=max(1, radius//6)))
     
-    # Blend modes for more intense effect
-    if random.random() < 0.5:  # 50% chance for additive blending
+    if random.random() < 0.5:
         return Image.blend(image, overlay, 0.7)
     else:
         return Image.alpha_composite(image, overlay)
 
 def apply_hue_shift(image):
     """Apply random hue shift to card pixels only"""
-    if random.random() > 0.6:  # 40% chance to apply hue shift
+    if random.random() > 0.6:
         return image
 
     img_array = np.array(image)
     rgb = img_array[..., :3]
     alpha = img_array[..., 3] if img_array.shape[2] == 4 else None
 
-    # Create mask of opaque pixels
     mask = (alpha > 0) if alpha is not None else np.ones(rgb.shape[:2], bool)
-
-    # Convert to HSV (uint8 format)
     hsv = cv2.cvtColor(rgb, cv2.COLOR_RGB2HSV)
     h, s, v = cv2.split(hsv)
 
-    # Apply hue shift only to masked pixels
     hue_shift = random.randint(-15, 15)
     h_float = h.astype(np.float32)
     h_float[mask] = (h_float[mask] + hue_shift) % 180.0
     h = np.clip(h_float, 0, 179).astype(np.uint8)
 
-    # Apply saturation adjustment to same pixels
     if random.random() < 0.3:
         sat_shift = random.uniform(0.7, 1.3)
         s_float = s.astype(np.float32)
         s_float[mask] = np.clip(s_float[mask] * sat_shift, 0, 255)
         s = s_float.astype(np.uint8)
 
-    # Merge channels
     hsv_shifted = cv2.merge((h, s, v))
     rgb_shifted = cv2.cvtColor(hsv_shifted, cv2.COLOR_HSV2RGB)
 
-    # Reconstruct image
     if alpha is not None:
         return Image.fromarray(np.dstack((rgb_shifted, alpha)), 'RGBA')
     return Image.fromarray(rgb_shifted, 'RGB')
 
 def apply_lighting_effects(image):
     """Apply brightness/contrast adjustments"""
-    if random.random() > 0.5:  # 50% chance to apply
+    if random.random() > 0.5:
         return image
     
-    # Random brightness adjustment
     brightness = random.uniform(0.8, 1.2)
     enhancer = ImageEnhance.Brightness(image)
     image = enhancer.enhance(brightness)
     
-    # Random contrast adjustment
     if random.random() < 0.3:
         contrast = random.uniform(0.9, 1.5)
         enhancer = ImageEnhance.Contrast(image)
@@ -153,30 +144,25 @@ def apply_lighting_effects(image):
     return image
 
 def process_card_image(card_image):
-    # Apply all visual effects first
     card_image = apply_glare_effect(card_image)
     card_image = apply_hue_shift(card_image)
     card_image = apply_lighting_effects(card_image)
 
     original_width, original_height = card_image.size
     
-    # Apply rotation
     angle = random.uniform(-179, 179)
     rotated = card_image.rotate(angle, expand=True, resample=Image.BICUBIC)
     
-    # Apply resizing
     new_width = random.randint(100, 300)
     scale_factor = new_width / rotated.width
     new_height = int(rotated.height * scale_factor)
     resized = rotated.resize((new_width, new_height), Image.LANCZOS)
     
-    # Create canvas for perspective transform
     padding = int(max(new_width, new_height) * 0.3)
     canvas_size = (new_width + padding*2, new_height + padding*2)
     canvas = Image.new('RGBA', canvas_size, (0, 0, 0, 0))
     canvas.paste(resized, (padding, padding))
     
-    # Perspective transform points
     src_points = np.array([
         [padding, padding],
         [padding + new_width, padding],
@@ -231,7 +217,6 @@ def process_card_image(card_image):
     except Exception:
         transformed = canvas
     
-    # Crop to visible area
     bbox = transformed.getbbox()
     if bbox:
         transformed = transformed.crop(bbox)
@@ -295,13 +280,11 @@ def generate_composite(composite_id, cards, backgrounds):
                 
             tw, th = transformed.size
             
-            # Calculate maximum allowed position to keep full card visible
             max_x = IMAGE_SIZE - tw
             max_y = IMAGE_SIZE - th
             if max_x < 0 or max_y < 0:
-                continue  # Skip if card is larger than image
+                continue
 
-            # Try to find non-overlapping position
             max_attempts = 10
             placed = False
             for _ in range(max_attempts):
@@ -309,15 +292,12 @@ def generate_composite(composite_id, cards, backgrounds):
                 y = random.randint(0, max_y)
                 new_bbox = (x, y, x + tw, y + th)
                 
-                # Check for complete overlap with existing cards
                 complete_overlap = False
                 for existing_bbox in annotations:
-                    # Check if new is completely inside existing
                     inside_existing = (new_bbox[0] >= existing_bbox[0] and
                                       new_bbox[1] >= existing_bbox[1] and
                                       new_bbox[2] <= existing_bbox[2] and
                                       new_bbox[3] <= existing_bbox[3])
-                    # Check if existing is completely inside new
                     existing_inside_new = (existing_bbox[0] >= new_bbox[0] and
                                           existing_bbox[1] >= new_bbox[1] and
                                           existing_bbox[2] <= new_bbox[2] and
@@ -332,12 +312,9 @@ def generate_composite(composite_id, cards, backgrounds):
                     break
 
             if not placed:
-                continue  # Skip this card after max attempts
+                continue
             
-            # Create axis-aligned bounding box
             bbox = new_bbox
-            
-            # Paste card onto background
             bg.paste(transformed, (x, y), transformed)
             annotations.append(bbox)
 
@@ -345,13 +322,10 @@ def generate_composite(composite_id, cards, backgrounds):
             print(f"Error processing card: {e}")
 
     if annotations:
-        # Save image
         bg.convert('RGB').save(f"{OUTPUT_DIR}/yolo/images/train/{composite_id}.jpg")
         
-        # Save YOLO annotations
         with open(f"{OUTPUT_DIR}/yolo/labels/train/{composite_id}.txt", "w") as f:
             for x_min, y_min, x_max, y_max in annotations:
-                # Calculate normalized center coordinates and dimensions
                 width = x_max - x_min
                 height = y_max - y_min
                 cx = (x_min + x_max) / 2 / IMAGE_SIZE
@@ -359,13 +333,11 @@ def generate_composite(composite_id, cards, backgrounds):
                 w = width / IMAGE_SIZE
                 h = height / IMAGE_SIZE
                 
-                # Ensure values are within valid range [0, 1]
                 cx = max(0.0, min(cx, 1.0))
                 cy = max(0.0, min(cy, 1.0))
                 w = max(0.0, min(w, 1.0))
                 h = max(0.0, min(h, 1.0))
                 
-                # Skip boxes that are too small
                 if w * h < (100 / (IMAGE_SIZE**2)):
                     continue
                 
@@ -381,8 +353,8 @@ def load_backgrounds():
             print(f"Skipping invalid background: {bg_path}")
     return backgrounds
 
-def create_dataset(csv_path):
-    cards = load_cards(csv_path)
+def create_dataset(csv_paths):
+    cards = load_cards(csv_paths)
     backgrounds = load_backgrounds()
     
     if not backgrounds:
@@ -392,5 +364,18 @@ def create_dataset(csv_path):
         generate_composite(composite_id, cards, backgrounds)
 
 if __name__ == "__main__":
-    create_dataset(r"$magic-{table}_202504101030.csv")
+    # Example usage:
+    # Single CSV file
+    # create_dataset("cards.csv")
+    
+    # Multiple CSV files
+    create_dataset([
+        r"$magic-{table}_202504101030.csv",
+        r"$Pokemon-{table}_202504112100.csv",
+        r"$YuGiOh-{table}_202504112056.csv"
+    ])
+    
+    # Or using glob pattern
+    # create_dataset(glob.glob("card_data/*.csv"))
+    
     print("Dataset generation complete!")
